+++ src/main/java/pl/clockworkjava/domain/guest/dto/GuestDTO.java
@@ -2,13 +2,13 @@ package pl.clockworkjava.domain.guest.dto;
 
 public class GuestDTO {
 
-    private final int id;
+    private final long id;
     private final String firstName;
     private final String lastName;
     private final int age;
     private final String gender;
 
-    public GuestDTO(int id, String firstName, String lastName, int age, String gender) {
+    public GuestDTO(long id, String firstName, String lastName, int age, String gender) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -16,7 +16,7 @@ public class GuestDTO {
         this.gender = gender;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
