+++ src/main/java/pl/clockworkjava/domain/reservation/Reservation.java
@@ -39,7 +39,7 @@ public class Reservation {
     public ReservationDTO getAsDTO() {
         return new ReservationDTO(this.id, this.from,
                 this.to, (int)this.room.getId(),
-                this.room.getNumber(), this.guest.getId(),
+                this.room.getNumber(), (int)this.guest.getId(),
                 this.guest.getFirstName() + " " + this.guest.getLastName());
     }
 }
